# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseCard": {
        "description": "A write Case as the chat renders it. Deliberately thin: ids, the lifecycle status,\nthe human-readable one-liner, and whether THIS caller is the one who must decide. The\nbusiness fields ride in ``detail`` untyped because they differ per agent (dates vs\nproject vs tool list) \u2014 the card shows them, it does not reason about them.\n\n``can_decide`` is a UI courtesy ONLY. The decision route re-checks server-side that the\ncaller is the Case's approver; a client that flips this flag gains nothing.",
        "properties": {
          "agent": {
            "description": "leave | jira | onboarding.",
            "title": "Agent",
            "type": "string"
          },
          "approver_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who must decide (null when there is nothing to decide).",
            "title": "Approver Email"
          },
          "can_decide": {
            "default": false,
            "description": "True when the CALLER is this Case's approver. UI hint only.",
            "title": "Can Decide",
            "type": "boolean"
          },
          "case_id": {
            "description": "The Case's id (the unit of work, not the message).",
            "title": "Case Id",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "description": "Agent-specific business fields.",
            "title": "Detail",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why it was denied or could not be routed, when it was.",
            "title": "Reason"
          },
          "requester_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who filed it.",
            "title": "Requester Email"
          },
          "status": {
            "description": "Lifecycle status, e.g. pending_approval / booked / dead_letter.",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "default": "",
            "description": "One-line human summary of what was requested.",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "agent",
          "status"
        ],
        "title": "CaseCard",
        "type": "object"
      },
      "ChatRequest": {
        "description": "Body for ``POST /chat``.",
        "properties": {
          "message": {
            "description": "The user's question.",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "default": "demo",
            "description": "Conversation id; each session keeps its own history.",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "description": "Standardized envelope for ``POST /chat``.\n\n``status`` lets the client distinguish a grounded answer from a graceful\nnon-answer without string-matching the prose: ``ok`` (answered), ``partial``\n(a real but incomplete answer, e.g. policy-only when a tool was down),\n``no_results`` (nothing found), ``blocked`` (RBAC-restricted), ``refused``\n(out of scope), ``tool_unavailable`` (the specialist's read tool could not run).",
        "properties": {
          "answer": {
            "description": "The assistant's answer text.",
            "title": "Answer",
            "type": "string"
          },
          "cases": {
            "description": "Write Cases this turn produced or is asking a decision on. Empty on every read turn. Rendered as cards; the approve/deny buttons act on them.",
            "items": {
              "$ref": "#/components/schemas/CaseCard"
            },
            "title": "Cases",
            "type": "array"
          },
          "latency_ms": {
            "description": "Server-side time to produce the answer.",
            "title": "Latency Ms",
            "type": "integer"
          },
          "session_id": {
            "description": "The conversation id this reply belongs to.",
            "title": "Session Id",
            "type": "string"
          },
          "sources": {
            "description": "Cited documents behind the answer (empty when not grounded).",
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "description": "ok | partial | no_results | blocked | refused | tool_unavailable.",
            "enum": [
              "ok",
              "partial",
              "no_results",
              "blocked",
              "refused",
              "tool_unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "answer",
          "latency_ms",
          "session_id",
          "status"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "CreateSessionResponse": {
        "description": "Response for ``POST /chat/sessions``.",
        "properties": {
          "session_id": {
            "description": "The newly created conversation id.",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "CreateSessionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryResponse": {
        "description": "Response for ``GET /chat/history/{session_id}``.",
        "properties": {
          "history": {
            "description": "Ordered {role, content} messages.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "History",
            "type": "array"
          },
          "session_id": {
            "description": "The session the history belongs to.",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "MessageResponse": {
        "description": "Generic ``{\"message\": ...}`` acknowledgement.",
        "properties": {
          "message": {
            "description": "Human-readable result.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "RenameSessionRequest": {
        "description": "Body for ``PATCH /chat/sessions/{session_id}``.",
        "properties": {
          "title": {
            "description": "New session title.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "RenameSessionRequest",
        "type": "object"
      },
      "SessionInfo": {
        "description": "One conversation owned by the caller (list item for the sidebar).",
        "properties": {
          "session_id": {
            "description": "Conversation id.",
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "User-set title, if any.",
            "title": "Title"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last activity (ISO 8601).",
            "title": "Updated At"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "SessionInfo",
        "type": "object"
      },
      "Source": {
        "description": "One cited chunk, in the frozen shape the React client binds to.",
        "properties": {
          "document_id": {
            "description": "Stable doc id = path relative to docs/ (e.g. hr/x.md).",
            "title": "Document Id",
            "type": "string"
          },
          "file": {
            "description": "Human-facing filename (basename of document_id).",
            "title": "File",
            "type": "string"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Parent section heading, when known.",
            "title": "Section"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Access tier the chunk came from (all/manager/hr_only).",
            "title": "Source Type"
          }
        },
        "required": [
          "document_id",
          "file"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered chatbot backend using LiteLLM + RAG",
    "title": "Company Chatbot API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat": {
      "post": {
        "description": "Send a message and receive an AI response with RAG context.\n\nThe user's role gates which document tiers are retrievable (RBAC).",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/history/{session_id}": {
      "delete": {
        "description": "Clear conversation history for one of the caller's own sessions.",
        "operationId": "clear_history_chat_history__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear History",
        "tags": [
          "Chat"
        ]
      },
      "get": {
        "description": "Retrieve conversation history for one of the caller's own sessions.",
        "operationId": "get_history_chat_history__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get History",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/sessions": {
      "get": {
        "description": "List the caller's own conversations, most recently active first.",
        "operationId": "list_sessions_chat_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  },
                  "title": "Response List Sessions Chat Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "description": "Create a new empty conversation owned by the caller.",
        "operationId": "create_session_chat_sessions_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/sessions/{session_id}": {
      "delete": {
        "description": "Delete one of the caller's conversations and its messages.",
        "operationId": "delete_session_chat_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Session",
        "tags": [
          "Chat"
        ]
      },
      "patch": {
        "description": "Rename one of the caller's conversations.",
        "operationId": "rename_session_chat_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rename Session",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "description": "Stream the answer token-by-token over SSE.\n\nSame auth/RBAC as ``POST /chat``; the body is identical. Emits typed events\n(token/sources/done/error). The service yields event dicts with a structured\n``data`` payload; here we JSON-encode that payload onto the SSE ``data`` field.",
        "operationId": "chat_stream_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat Stream",
        "tags": [
          "Chat"
        ]
      }
    },
...
  }
}